+++ experimental/packages/opentelemetry-instrumentation-http/src/http.ts
@@ -1,18 +1,8 @@
 /*
  * Copyright The OpenTelemetry Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * SPDX-License-Identifier: Apache-2.0
  */
+
 import {
   context,
   HrTime,
@@ -91,6 +81,8 @@ export class HttpInstrumentation extends InstrumentationBase<HttpInstrumentation
   /** keep track on spans not ended */
   private readonly _spanNotEnded: WeakSet<Span> = new WeakSet<Span>();
   private _headerCapture;
+  private _httpPatched: boolean = false;
+  private _httpsPatched: boolean = false;
   declare private _oldHttpServerDurationHistogram: Histogram;
   declare private _stableHttpServerDurationHistogram: Histogram;
   declare private _oldHttpClientDurationHistogram: Histogram;
@@ -209,8 +201,16 @@ export class HttpInstrumentation extends InstrumentationBase<HttpInstrumentation
       'http',
       ['*'],
       (moduleExports: Http): Http => {
+        // Guard against double-instrumentation, if loaded by both `require`
+        // and `import`.
+        if (this._httpPatched) {
+          return moduleExports;
+        }
+        this._httpPatched = true;
+
         // eslint-disable-next-line @typescript-eslint/no-explicit-any
         const isESM = (moduleExports as any)[Symbol.toStringTag] === 'Module';
+
         if (!this.getConfig().disableOutgoingRequestInstrumentation) {
           const patchedRequest = this._wrap(
             moduleExports,
@@ -241,6 +241,7 @@ export class HttpInstrumentation extends InstrumentationBase<HttpInstrumentation
         return moduleExports;
       },
       (moduleExports: Http) => {
+        this._httpPatched = false;
         if (moduleExports === undefined) return;
 
         if (!this.getConfig().disableOutgoingRequestInstrumentation) {
@@ -259,8 +260,16 @@ export class HttpInstrumentation extends InstrumentationBase<HttpInstrumentation
       'https',
       ['*'],
       (moduleExports: Https): Https => {
+        // Guard against double-instrumentation, if loaded by both `require`
+        // and `import`.
+        if (this._httpsPatched) {
+          return moduleExports;
+        }
+        this._httpsPatched = true;
+
         // eslint-disable-next-line @typescript-eslint/no-explicit-any
         const isESM = (moduleExports as any)[Symbol.toStringTag] === 'Module';
+
         if (!this.getConfig().disableOutgoingRequestInstrumentation) {
           const patchedRequest = this._wrap(
             moduleExports,
@@ -291,6 +300,7 @@ export class HttpInstrumentation extends InstrumentationBase<HttpInstrumentation
         return moduleExports;
       },
       (moduleExports: Https) => {
+        this._httpsPatched = false;
         if (moduleExports === undefined) return;
 
         if (!this.getConfig().disableOutgoingRequestInstrumentation) {
@@ -459,12 +469,15 @@ export class HttpInstrumentation extends InstrumentationBase<HttpInstrumentation
           this._callResponseHook(span, response);
         }
 
-        this._headerCapture.client.captureRequestHeaders(span, header =>
-          request.getHeader(header)
+        span.setAttributes(
+          this._headerCapture.client.captureRequestHeaders(header =>
+            request.getHeader(header)
+          )
         );
-        this._headerCapture.client.captureResponseHeaders(
-          span,
-          header => response.headers[header]
+        span.setAttributes(
+          this._headerCapture.client.captureResponseHeaders(
+            header => response.headers[header]
+          )
         );
 
         context.bind(context.active(), response);
@@ -623,6 +636,13 @@ export class HttpInstrumentation extends InstrumentationBase<HttpInstrumentation
         instrumentation._diag
       );
 
+      Object.assign(
+        spanAttributes,
+        instrumentation._headerCapture.server.captureRequestHeaders(
+          header => request.headers[header]
+        )
+      );
+
       const spanOptions: SpanOptions = {
         kind: SpanKind.SERVER,
         attributes: spanAttributes,
@@ -663,11 +683,6 @@ export class HttpInstrumentation extends InstrumentationBase<HttpInstrumentation
           if (instrumentation.getConfig().responseHook) {
             instrumentation._callResponseHook(span, response);
           }
-
-          instrumentation._headerCapture.server.captureRequestHeaders(
-            span,
-            header => request.headers[header]
-          );
 
           // After 'error', no further events other than 'close' should be emitted.
           let hasError = false;
@@ -889,8 +904,10 @@ export class HttpInstrumentation extends InstrumentationBase<HttpInstrumentation
       getIncomingStableRequestMetricAttributesOnResponse(attributes)
     );
 
-    this._headerCapture.server.captureResponseHeaders(span, header =>
-      response.getHeader(header)
+    span.setAttributes(
+      this._headerCapture.server.captureResponseHeaders(header =>
+        response.getHeader(header)
+      )
     );
 
     span.setAttributes(attributes).setStatus({