+++ experimental/packages/opentelemetry-instrumentation-http/src/utils.ts
@@ -1,17 +1,6 @@
 /*
  * Copyright The OpenTelemetry Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * SPDX-License-Identifier: Apache-2.0
  */
 import {
   Attributes,
@@ -1108,9 +1097,9 @@ export function headerCapture(
   }
 
   return (
-    span: Span,
     getHeader: (key: string) => undefined | string | string[] | number
-  ) => {
+  ): Attributes => {
+    const attributes: Attributes = {};
     for (const capturedHeader of normalizedHeaders.keys()) {
       const value = getHeader(capturedHeader);
 
@@ -1122,13 +1111,14 @@ export function headerCapture(
       const key = `http.${type}.header.${normalizedHeader}`;
 
       if (typeof value === 'string') {
-        span.setAttribute(key, [value]);
+        attributes[key] = [value];
       } else if (Array.isArray(value)) {
-        span.setAttribute(key, value);
+        attributes[key] = value;
       } else {
-        span.setAttribute(key, [value]);
+        attributes[key] = [value];
       }
     }
+    return attributes;
   };
 }